webapp/requirements.txt
# Wymagania dla WebSocket Client (Python)
# Instalacja: pip install -r requirements.txt

websocket-client>=1.0.0
requests>=2.31.0
